feat(reviews): per-record-version data model changes
