+++ invenio_rdm_records/records/mappings/os-v2/rdmrecords/drafts/draft-v6.0.0.json
@@ -1643,6 +1643,49 @@
             }
           }
         }
+      },
+      "review": {
+        "type": "object",
+        "properties": {
+          "$schema": {
+            "type": "keyword",
+            "index": false
+          },
+          "id": {
+            "type": "keyword"
+          },
+          "type": {
+            "type": "keyword"
+          },
+          "title": {
+            "type": "text"
+          },
+          "description": {
+            "type": "text"
+          },
+          "status": {
+            "type": "keyword"
+          },
+          "payload": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "topic": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "receiver": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "created_by": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "@v": {
+            "type": "keyword"
+          }
+        }
       }
     }
   }
