+++ invenio_rdm_records/records/mappings/v7/rdmrecords/drafts/draft-v6.0.0.json
@@ -1581,6 +1581,49 @@
             "type": "boolean"
           }
         }
+      },
+      "review": {
+        "type": "object",
+        "properties": {
+          "$schema": {
+            "type": "keyword",
+            "index": false
+          },
+          "id": {
+            "type": "keyword"
+          },
+          "type": {
+            "type": "keyword"
+          },
+          "title": {
+            "type": "text"
+          },
+          "description": {
+            "type": "text"
+          },
+          "status": {
+            "type": "keyword"
+          },
+          "payload": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "topic": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "receiver": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "created_by": {
+            "type": "object",
+            "dynamic": "true"
+          },
+          "@v": {
+            "type": "keyword"
+          }
+        }
       }
     }
   }
